// src/api/models/userSocialAccountResDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { UserSocialAccountResDTOSocialEmail } from './userSocialAccountResDTOSocialEmail';
import type { UserSocialAccountResDTOSocialType } from './userSocialAccountResDTOSocialType';

export interface UserSocialAccountResDTO {
  /** 소셜 계정 타입 */
  socialType: UserSocialAccountResDTOSocialType;
  /**
   * 소셜 계정 이메일
   * @nullable
   */
  socialEmail: UserSocialAccountResDTOSocialEmail;
}