// src/api/models/userSocialAccountResDTOSocialType.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */

/**
 * 소셜 계정 타입
 */
export type UserSocialAccountResDTOSocialType = typeof UserSocialAccountResDTOSocialType[keyof typeof UserSocialAccountResDTOSocialType];


export const UserSocialAccountResDTOSocialType = {
  KAKAO: 'KAKAO',
  NAVER: 'NAVER',
  GOOGLE: 'GOOGLE',
  APPLE: 'APPLE',
} as const;